Moving hdfs functions
